+++ Celeste_Launcher_Gui/Config.cs
@@ -33,8 +33,8 @@ namespace Celeste_Launcher_Gui
         [XmlElement(ElementName = "BetaUpdate")]
         public bool BetaUpdate { get; set; }
 
-        [XmlElement(ElementName = "IsSteam")]
-        public bool IsSteam { get; set; }
+        [XmlIgnore]
+        public bool IsSteamVersion { get; set; } = false;
 
         [XmlElement(ElementName = "LoginInfo")]
         public LoginInfo LoginInfo { get; set; } = new LoginInfo();